TDD fail: tests pass even though functionality not implemented
Trying to verify the requirement that all rooms are reachable
from a random starting point has shown that the dungeon generator
class needs to be split up. The existing tests aren't granular
enough to show what causes a failure, and they are complicated
enough that they can pass even without the functionality that
should be required to pass.

I'll use this as a chance to refactor into smaller units that can
be tested independently.